RE-39: Переоформил былой функционал в микросервис
