+++ internal/banner-service/controller/http/bannerRouter.go
@@ -1,0 +1,25 @@
+package banner
+
+import (
+	"net/http"
+	banner "retarget/internal/banner-service/usecase"
+
+	"github.com/gorilla/mux"
+)
+
+type BannerController struct {
+	BannerUsecase *banner.BannerUsecase
+}
+
+func NewBannerController(bannerUsecase *banner.BannerUsecase) *BannerController {
+	return &BannerController{BannerUsecase: bannerUsecase}
+}
+
+func SetupBannerRoutes(bannerUsecase *banner.BannerUsecase) http.Handler {
+	muxRouter := mux.NewRouter()
+	bannerController := NewBannerController(bannerUsecase)
+	// middleware.AuthMiddleware(authUsecase)()
+	muxRouter.Handle("/api/v1/banner/user/{user_id}/all", http.HandlerFunc(bannerController.GetBannersByUserCookie))
+
+	return muxRouter
+}
