+++ internal/banner-service/entity/banner.go
